phase 1: refactoring
