+++ pkg/omsecrets/openmetadata_secrets_reader.go
@@ -5,10 +5,12 @@ package omsecrets
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/hashicorp/go-hclog"
 
 	omclient "fybrik.io/vault-plugin-secrets-omd-reader/pkg/openmetadata"
+	dbtypes "fybrik.io/vault-plugin-secrets-omd-reader/pkg/openmetadata/database-types"
 )
 
 type OpenMetadataSecretsReader struct {
@@ -17,10 +19,19 @@ type OpenMetadataSecretsReader struct {
 
 // GetSecret returns the content of openmetadata secret.
 func (s *OpenMetadataSecretsReader) GetSecret(ctx context.Context, secretName string, log hclog.Logger) (map[string]interface{}, error) {
+	nameToDatabaseStruct := map[string]dbtypes.Databasetype{
+		"Datalake": dbtypes.S3{},
+	}
+
 	databaseService, err := s.client.GetConnectionInformation(ctx, secretName, log)
 	if err != nil {
 		return nil, err
 	}
+	serviceType := databaseService.GetServiceType()
+	dt, found := nameToDatabaseStruct[serviceType]
+	if !found {
+		return nil, fmt.Errorf("Service type %s not recognized", serviceType)
+	}
 	config := databaseService.Connection.GetConfig()
-	return s.client.ExtractSecretsFromConfig(config)
+	return dt.ExtractSecretsFromConfig(config)
 }
